Added Functions For Arena.vb Movement.
clearCoordinates function clears every imagebox inside the function. (needs to be updated for your map)

showPlayerCoordinate is the primary movement function that calls the clearCoordinates function, then makes the player's respective imagebox show the player sprite according to the player's x and y variable.

The functions have yet to be adapted to your current program.